requirements_custom.txt
python-dotenv
openai
streamlit
numpy
